refactoring to allow new modules without confusion
